pyproject.toml
[project]
name = "gradient-technology"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "pytest-asyncio>=1.3.0",
    "pytest[asyncio]>=9.0.2",
    "python-dotenv>=1.2.2",
    "redis>=7.2.1",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

